+++ src/main/kotlin/com/kraktun/kbot/commands/groups/LockCommand.kt
@@ -1,4 +1,4 @@
-package com.kraktun.kbot.commands
+package com.kraktun.kbot.commands.groups
 
 import com.kraktun.kbot.commands.core.BaseCommand
 import com.kraktun.kbot.commands.core.ChatOptions
